refactor MatchPattern
