+++ common/addons/config-locators/src/main/java/com/dfsek/terra/addons/feature/locator/config/pattern/AirMatchPatternTemplate.java
@@ -3,8 +3,7 @@ package com.dfsek.terra.addons.feature.locator.config.pattern;
 import com.dfsek.tectonic.annotations.Value;
 import com.dfsek.tectonic.loading.object.ObjectTemplate;
 import com.dfsek.terra.addons.feature.locator.patterns.Pattern;
-import com.dfsek.terra.addons.feature.locator.patterns.match.MatchPattern;
-import com.dfsek.terra.api.TerraPlugin;
+import com.dfsek.terra.addons.feature.locator.patterns.MatchPattern;
 import com.dfsek.terra.api.block.state.BlockState;
 import com.dfsek.terra.api.config.meta.Meta;
 import com.dfsek.terra.api.util.Range;
