+++ common/addons/config-locators/src/main/java/com/dfsek/terra/addons/feature/locator/config/pattern/BlockSetMatchPatternTemplate.java
@@ -3,7 +3,7 @@ package com.dfsek.terra.addons.feature.locator.config.pattern;
 import com.dfsek.tectonic.annotations.Value;
 import com.dfsek.tectonic.loading.object.ObjectTemplate;
 import com.dfsek.terra.addons.feature.locator.patterns.Pattern;
-import com.dfsek.terra.addons.feature.locator.patterns.match.MatchPattern;
+import com.dfsek.terra.addons.feature.locator.patterns.MatchPattern;
 import com.dfsek.terra.api.config.meta.Meta;
 import com.dfsek.terra.api.util.Range;
 import com.dfsek.terra.api.util.collection.MaterialSet;
