+++ common/addons/config-locators/src/main/java/com/dfsek/terra/addons/feature/locator/patterns/MatchPattern.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.addons.feature.locator.patterns.match;
+package com.dfsek.terra.addons.feature.locator.patterns;
 
 import com.dfsek.terra.addons.feature.locator.patterns.Pattern;
 import com.dfsek.terra.api.block.state.BlockState;
